# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GenericAgent Admin API",
    "version": "0.3.4"
  },
  "paths": {
...
    "/api/conductor/settings": {
      "post": {
        "summary": "Update Conductor Settings",
        "description": "Update conductor automation policy (currently: auto-accept).",
        "operationId": "update_conductor_settings_api_conductor_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConductorSettingsReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorStatusResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conductor/readme/{topic}": {
      "get": {
        "summary": "Get Readme Topic",
        "description": "Return one conductor README topic (OpenAPI-documented).",
        "operationId": "get_readme_topic_api_conductor_readme__topic__get",
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorTextResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conductor/chat": {
      "get": {
        "summary": "Get Chat",
        "description": "Return the last N conductor chat messages for hydration.",
        "operationId": "get_chat_api_conductor_chat_get",
        "parameters": [
          {
            "name": "last",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Last"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorChatListResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Chat",
        "description": "Admit one chat message for the conductor and broadcast it.\n\nA user message whose ``request_id`` names an open workflow appends to\nthat workflow (conversation continuity); unknown or closed ids fall back\nto admitting a fresh task.",
        "operationId": "post_chat_api_conductor_chat_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConductorChatIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorChatMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conductor/subagent": {
      "get": {
        "summary": "List Subagents",
        "description": "Return the subagent pool snapshot the UI renders.",
        "operationId": "list_subagents_api_conductor_subagent_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSubagentListResp"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Start Subagent",
        "description": "Dispatch a new subagent through the service policy boundary.",
        "operationId": "start_subagent_api_conductor_subagent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConductorStartSubagent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSubagentInstructionResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conductor/operations/{operation_id}": {
      "get": {
        "summary": "Get Operation",
        "operationId": "get_operation_api_conductor_operations__operation_id__get",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorOperationResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conductor/workflow": {
      "get": {
        "summary": "List Workflows",
        "description": "Return the workflow tracker snapshot for the UI board.",
        "operationId": "list_workflows_api_conductor_workflow_get",
        "parameters": [
          {
            "name": "last",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Last"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorWorkflowListResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conductor/subagent/{sid}": {
      "get": {
        "summary": "Get Subagent",
        "description": "Full worker dossier for human review (engine reply + mirror facts).",
        "operationId": "get_subagent_api_conductor_subagent__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          },
          {
            "name": "max_len",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 1,
              "default": 5000,
              "title": "Max Len"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSubagent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Subagent Action",
        "description": "Apply one verb (keyinfo/accept/rework/input/abort/...) to a worker.",
        "operationId": "subagent_action_api_conductor_subagent__sid__post",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConductorSubagentAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSubagentActionResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conductor/journal": {
      "get": {
        "summary": "Get Conductor Journal",
        "description": "Catch-up read of the engine's durable run journal (P2-A).\n\nSSE stays the live hint; clients that dropped events re-sync from the\njournal instead of guessing.  Disabled journals pass through the\nengine's explicit `disabled` payload unchanged.",
        "operationId": "get_conductor_journal_api_conductor_journal_get",
        "parameters": [
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Seq"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conductor/start": {
      "post": {
        "summary": "Start Conductor",
        "description": "Start the conductor supervisor.",
        "operationId": "start_conductor_api_conductor_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ConductorStartReq"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorLifecycleResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConductorChatIn": {
        "properties": {
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "role": {
            "type": "string",
            "enum": [
              "conductor",
              "system",
              "user"
            ],
            "title": "Role",
            "default": "conductor"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "final": {
            "type": "boolean",
            "title": "Final",
            "default": false
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "llm_index": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Index"
          },
          "subagent_llm_index": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Subagent Llm Index"
          },
          "subagent_model_policy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "follow_main",
                  "default",
                  "locked"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Subagent Model Policy"
          }
        },
        "type": "object",
        "required": [
          "msg"
        ],
        "title": "ConductorChatIn"
      },
      "ConductorChatListResp": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConductorChatMessage"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConductorChatListResp"
      },
      "ConductorChatMessage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "ts": {
            "type": "integer",
            "title": "Ts"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "final",
                  "error"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "final": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "role",
          "msg",
          "ts"
        ],
        "title": "ConductorChatMessage"
      },
...
      "ConductorSettingsReq": {
        "properties": {
          "auto_accept": {
            "type": "boolean",
            "title": "Auto Accept"
          }
        },
        "type": "object",
        "required": [
          "auto_accept"
        ],
        "title": "ConductorSettingsReq"
      },
...
      "ConductorStatusResp": {
        "properties": {
          "started": {
            "type": "boolean",
            "title": "Started"
          },
          "stopping": {
            "type": "boolean",
            "title": "Stopping"
          },
          "admission_open": {
            "type": "boolean",
            "title": "Admission Open"
          },
          "loop_alive": {
            "type": "boolean",
            "title": "Loop Alive"
          },
          "agent_alive": {
            "type": "boolean",
            "title": "Agent Alive"
          },
          "subagents": {
            "$ref": "#/components/schemas/ConductorSubagentCounts"
          },
          "chat_count": {
            "type": "integer",
            "title": "Chat Count"
          },
          "auto_accept": {
            "type": "boolean",
            "title": "Auto Accept",
            "default": true
          },
          "boot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boot Id"
          },
          "protocol_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Version"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "path_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Policy"
          },
          "recovery": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery"
          }
        },
        "type": "object",
        "required": [
          "started",
          "stopping",
          "admission_open",
          "loop_alive",
          "agent_alive",
          "subagents",
          "chat_count"
        ],
        "title": "ConductorStatusResp"
      },
...
      "ConductorSubagentInstructionResp": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ConductorSubagentInstructionResp"
      },
      "ConductorSubagentListResp": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConductorSubagent"
            },
            "type": "array",
            "title": "Items"
          },
          "boot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boot Id"
          },
          "snapshot_revision": {
            "type": "integer",
            "title": "Snapshot Revision",
            "default": -1
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConductorSubagentListResp"
      },
      "ConductorTextResp": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ConductorTextResp"
      },
...
      "ConductorWorkflowListResp": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConductorWorkflow"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConductorWorkflowListResp"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}